+++ src/constants.ts
@@ -19,16 +19,17 @@ export const DEFAULT_SETTINGS: ObsidianGitSettings = {
     autoBackupAfterFileChange: false,
     treeStructure: false,
     refreshSourceControl: true,
+    basePath: "",
 };
 
 export const GIT_VIEW_CONFIG = {
     type: 'git-view',
     name: 'Source Control',
-    icon: 'feather-git-pull-request'
+    icon: 'git-pull-request'
 };
 
 export const DIFF_VIEW_CONFIG = {
     type: 'diff-view',
     name: 'Diff View',
-    icon: 'feather-git-pull-request'
+    icon: 'git-pull-request'
 };