+++ src/statusBar.ts
@@ -51,37 +51,37 @@ export class StatusBar {
                 break;
             case PluginState.status:
                 this.statusBarEl.ariaLabel = "Checking repository status...";
-                setIcon(this.statusBarEl, "feather-refresh-cw");
+                setIcon(this.statusBarEl, "refresh-cw");
                 this.statusBarEl.addClass(this.base + "status");
                 break;
             case PluginState.add:
                 this.statusBarEl.ariaLabel = "Adding files...";
-                setIcon(this.statusBarEl, "feather-refresh-w");
+                setIcon(this.statusBarEl, "refresh-w");
                 this.statusBarEl.addClass(this.base + "add");
                 break;
             case PluginState.commit:
                 this.statusBarEl.ariaLabel = "Committing changes...";
-                setIcon(this.statusBarEl, "feather-git-commit");
+                setIcon(this.statusBarEl, "git-commit");
                 this.statusBarEl.addClass(this.base + "commit");
                 break;
             case PluginState.push:
                 this.statusBarEl.ariaLabel = "Pushing changes...";
-                setIcon(this.statusBarEl, "feather-upload");
+                setIcon(this.statusBarEl, "upload");
                 this.statusBarEl.addClass(this.base + "push");
                 break;
             case PluginState.pull:
                 this.statusBarEl.ariaLabel = "Pulling changes...";
-                setIcon(this.statusBarEl, "feather-download");
+                setIcon(this.statusBarEl, "download");
                 this.statusBarEl.addClass(this.base + "pull");
                 break;
             case PluginState.conflicted:
                 this.statusBarEl.ariaLabel = "You have conflict files...";
-                setIcon(this.statusBarEl, "feather-alert-circle");
+                setIcon(this.statusBarEl, "alert-circle");
                 this.statusBarEl.addClass(this.base + "conflict");
                 break;
             default:
                 this.statusBarEl.ariaLabel = "Failed on initialization!";
-                setIcon(this.statusBarEl, "feather-alert-triangle");
+                setIcon(this.statusBarEl, "alert-triangle");
                 this.statusBarEl.addClass(this.base + "failed-init");
                 break;
         }
@@ -91,11 +91,15 @@ export class StatusBar {
         if (timestamp) {
             const moment = (window as any).moment;
             const fromNow = moment(timestamp).fromNow();
-            this.statusBarEl.ariaLabel = `Last Git update: ${fromNow}`;
+            this.statusBarEl.ariaLabel = `${this.plugin.offlineMode ? "Offline: " : ""}Last Git update: ${fromNow}`;
         } else {
-            this.statusBarEl.ariaLabel = "Git is ready";
+            this.statusBarEl.ariaLabel = this.plugin.offlineMode ? "Git is offline" : "Git is ready";
         }
-        setIcon(this.statusBarEl, "feather-check");
+        if (this.plugin.offlineMode) {
+            setIcon(this.statusBarEl, "globe");
+        } else {
+            setIcon(this.statusBarEl, "check");
+        }
         this.statusBarEl.addClass(this.base + "idle");
     }
 }