+++ src/ui/diff/diffView.ts
@@ -1,35 +1,19 @@
 import { html } from "diff2html";
-import { ItemView, MarkdownView, WorkspaceLeaf } from "obsidian";
+import { ItemView, ViewStateResult, WorkspaceLeaf } from "obsidian";
 import { DIFF_VIEW_CONFIG } from "src/constants";
 import ObsidianGit from "src/main";
+import { DiffViewState } from "src/types";
 
 
 export default class DiffView extends ItemView {
-
-    filePath: string;
     parser: DOMParser;
-    intervalId: number;
     gettingDiff: boolean = false;
+    state: DiffViewState;
+
     constructor(leaf: WorkspaceLeaf, private plugin: ObsidianGit) {
         super(leaf);
         this.parser = new DOMParser();
-
-        this.registerEvent(this.app.workspace.on('active-leaf-change', (leaf) => {
-            if (leaf.view instanceof MarkdownView) {
-                this.filePath = leaf.view.file.path;
-            } else {
-                this.filePath = null;
-            }
-            this.refresh();
-        }));
-        addEventListener('diff-update', this.firstOpen.bind(this));
-        this.intervalId = window.setInterval(() => this.refresh(), 10000);
-        this.registerInterval(this.intervalId);
-    }
-
-    firstOpen(event: CustomEvent) {
-        this.filePath = event.detail.path;
-        this.refresh();
+        addEventListener('git-refresh', this.refresh.bind(this));
     }
 
     getViewType(): string {
@@ -44,9 +28,19 @@ export default class DiffView extends ItemView {
         return DIFF_VIEW_CONFIG.icon;
     }
 
+    async setState(state: any, result: ViewStateResult): Promise<void> {
+        this.state = state;
+
+        await this.refresh();
+        return;
+    }
+
+    getState() {
+        return this.state;
+    }
+
     onClose(): Promise<void> {
-        removeEventListener('diff-update', this.firstOpen.bind(this));
-        window.clearInterval(this.intervalId);
+        removeEventListener('git-refresh', this.refresh.bind(this));
         return super.onClose();
     }
 
@@ -56,30 +50,24 @@ export default class DiffView extends ItemView {
     }
 
     async refresh(): Promise<void> {
-        if (this.filePath && !this.gettingDiff) {
-            if (!this.app.vault.getAbstractFileByPath(this.filePath)) {
-                this.contentEl.empty();
+        if (this.state?.file && !this.gettingDiff && this.plugin.gitManager) {
+
+            this.gettingDiff = true;
+            const diff = this.parser.parseFromString(
+                html(await this.plugin.gitManager.getDiffString(this.state.file, this.state.staged)),
+                'text/html')
+                .querySelector('.d2h-file-diff');
+            this.contentEl.empty();
+            if (diff) {
+                this.contentEl.append(diff);
+            } else {
                 const div = this.contentEl.createDiv({ cls: 'diff-err' });
                 div.createSpan({ text: '⚠️', cls: 'diff-err-sign' });
                 div.createEl('br');
-                div.createSpan({ text: this.filePath + ' was deleted' });
-            } else {
-                this.gettingDiff = true;
-                const diff = this.parser.parseFromString(
-                    html(await this.plugin.gitManager.getDiffString(this.filePath)),
-                    'text/html')
-                    .querySelector('.d2h-file-diff');
-                this.contentEl.empty();
-                if (diff) {
-                    this.contentEl.append(diff);
-                } else {
-                    const div = this.contentEl.createDiv({ cls: 'diff-err' });
-                    div.createSpan({ text: '⚠️', cls: 'diff-err-sign' });
-                    div.createEl('br');
-                    div.createSpan({ text: 'No changes to' + this.filePath });
-                }
-                this.gettingDiff = false;
+                div.createSpan({ text: 'No changes to ' + this.state.file });
             }
+            this.gettingDiff = false;
+
         }
     }