+++ src/ui/sidebar/gitView.svelte
@@ -19,18 +19,22 @@
   let changesOpen = true;
   let stagedOpen = true;
   let loading = true;
-  const debRefresh = debounce(() => {
-    if (plugin.settings.refreshSourceControl) {
-      refresh();
-    }
-  }, 1000);
+  const debRefresh = debounce(
+    () => {
+      if (plugin.settings.refreshSourceControl) {
+        refresh();
+      }
+    },
+    7000,
+    true
+  );
 
   let showTree = plugin.settings.treeStructure;
   let layoutBtn: HTMLElement;
   $: {
     if (layoutBtn) {
       layoutBtn.empty();
-      setIcon(layoutBtn, showTree ? "feather-list" : "feather-folder", 16);
+      setIcon(layoutBtn, showTree ? "list" : "folder", 16);
     }
   }
 
@@ -39,14 +43,13 @@
   let createEvent: EventRef;
   let renameEvent: EventRef;
 
+  addEventListener("git-refresh", refresh);
   //This should go in the onMount callback, for some reason it doesn't fire though
   //setImmediate's callback will execute after the current event loop finishes.
   plugin.app.workspace.onLayoutReady(() =>
     setImmediate(() => {
       buttons.forEach((btn) => setIcon(btn, btn.getAttr("data-icon"), 16));
-      setIcon(layoutBtn, showTree ? "feather-list" : "feather-folder", 16);
-
-      refresh();
+      setIcon(layoutBtn, showTree ? "list" : "folder", 16);
 
       modifyEvent = plugin.app.vault.on("modify", () => {
         debRefresh();
@@ -61,8 +64,6 @@
         debRefresh();
       });
 
-      addEventListener('git-source-control-refresh', refresh);
-
       plugin.registerEvent(modifyEvent);
       plugin.registerEvent(deleteEvent);
       plugin.registerEvent(createEvent);
@@ -75,23 +76,20 @@
     plugin.app.metadataCache.offref(deleteEvent);
     plugin.app.metadataCache.offref(createEvent);
     plugin.app.metadataCache.offref(renameEvent);
-    removeEventListener('git-source-control-refresh', refresh);
-    
+    removeEventListener("git-refresh", refresh);
   });
 
   function commit() {
     loading = true;
-    plugin.gitManager.commit(commitMessage).then(() => {
-      if (commitMessage !== plugin.settings.commitMessage) {
-        commitMessage = "";
-      }
-      refresh();
-    });
+    plugin.gitManager
+      .commit(commitMessage)
+      .then(() => {
+        if (commitMessage !== plugin.settings.commitMessage) {
+          commitMessage = "";
+        }
+      })
+      .finally(refresh);
   }
-
-  addEventListener("git-refresh", (_) => {
-    refresh();
-  });
 
   async function refresh() {
     loading = true;
@@ -112,30 +110,22 @@
 
   function stageAll() {
     loading = true;
-    plugin.gitManager.stageAll().then(() => {
-      refresh();
-    });
+    plugin.gitManager.stageAll().finally(refresh);
   }
   function unstageAll() {
     loading = true;
-    plugin.gitManager.unstageAll().then(() => {
-      refresh();
-    });
+    plugin.gitManager.unstageAll().finally(refresh);
   }
   function push() {
     loading = true;
 
     if (ready) {
-      plugin.push().then((pushedFiles) => {
-        refresh();
-      });
+      plugin.push().finally(refresh);
     }
   }
   function pull() {
     loading = true;
-    plugin.pullChangesFromRemote().then(() => {
-      refresh();
-    });
+    plugin.pullChangesFromRemote().finally(refresh);
   }
 </script>
 
@@ -144,7 +134,7 @@
     <div class="group">
       <div
         id="commit-btn"
-        data-icon="feather-check"
+        data-icon="check"
         class="nav-action-button"
         aria-label="Commit"
         bind:this={buttons[0]}
@@ -153,7 +143,7 @@
       <div
         id="stage-all"
         class="nav-action-button"
-        data-icon="feather-plus-circle"
+        data-icon="plus-circle"
         aria-label="Stage all"
         bind:this={buttons[1]}
         on:click={stageAll}
@@ -161,7 +151,7 @@
       <div
         id="unstage-all"
         class="nav-action-button"
-        data-icon="feather-minus-circle"
+        data-icon="minus-circle"
         aria-label="Unstage all"
         bind:this={buttons[2]}
         on:click={unstageAll}
@@ -169,7 +159,7 @@
       <div
         id="push"
         class="nav-action-button"
-        data-icon="feather-upload"
+        data-icon="upload"
         aria-label="Push"
         bind:this={buttons[3]}
         on:click={push}
@@ -177,7 +167,7 @@
       <div
         id="pull"
         class="nav-action-button"
-        data-icon="feather-download"
+        data-icon="download"
         aria-label="Pull"
         bind:this={buttons[4]}
         on:click={pull}
@@ -198,7 +188,7 @@
       id="refresh"
       class="nav-action-button"
       class:loading
-      data-icon="feather-refresh-cw"
+      data-icon="refresh-cw"
       aria-label="Refresh"
       bind:this={buttons[6]}
       on:click={refresh}